fix(á»order): create order

+++ src/controllers/order.js
@@ -1,6 +1,7 @@
 const Order = require('../models/order');
 const OrderItem = require('../models/orderItem');
-const Cart = require('../models/cart');
+const OrderProduct = require('../models/orderProduct');
+const OrderAddress = require('../models/orderAddress');
 const CartItem = require('../models/cartItem');
 const Address = require('../models/address');
 const { successResponse, errorCommonResponse } = require('../utils').common;
@@ -16,7 +17,12 @@ controllers.getAllOrder = async (req, res) => {
             orders = await Order.create({ userId });
         }
         
-        const result = await OrderItem.find({ orderId: orders._id }).lean();
+        const result = await OrderItem
+            .find({ orderId: orders._id })
+            .populate('productItems')
+            .populate('receiver')
+            .lean();
+
         successResponse(res, result);
     } catch (exception) {
         errorCommonResponse(res, exception);
@@ -25,28 +31,41 @@ controllers.getAllOrder = async (req, res) => {
 
 controllers.createOrder = async (req, res) => {
     try {
-        const { _id: userId, orderId, cartId } = req.user;
+        const { orderId } = req.user;
         const { itemsCart, info } = req.body;
 
         const dataCart = await CartItem.find({ _id: { $in: itemsCart }}).populate('products').lean();
         const dataAddress = await Address.findById(info).select('name phone address -_id').lean();
 
-        const dataProduct = (dataCart || []).map(async (item) => {
-            const data = {
-                name: item.products.name,
-                price: item.products.price,
-                quantity: item.quantity,
-                avatar: item.products.avatar,
-                amount: item.amount,
-            };
-            return await OrderItem.create({...data, ...dataAddress, orderId});
-        });
+        if (!dataCart || !dataAddress) {
+            return errorCommonResponse(res, 'order fail');
+        }
 
-        const result = await Promise.all(dataProduct);
+        const dataProduct = (dataCart || []).map((item) => ({
+            name: item.products.name,
+            price: item.products.price,
+            quantity: item.quantity,
+            avatar: item.products.avatar,
+            amount: item.amount,
+        }));
+
+        const createProduct = await OrderProduct.insertMany(dataProduct);
+        const createAddress = await OrderAddress.create(dataAddress);
+        const productId = createProduct.map(product => product._id);
+        const total = createProduct.reduce((sum, product) => sum + product?.amount, 0);
+
+        if (!createProduct || !createAddress) {
+            return errorCommonResponse(res, 'order fail');
+        }
+
+        const result = await OrderItem.create({ total, productId, receiverId: createAddress._id, orderId });
+        if (!result) {
+            return errorCommonResponse(res, 'order fail');
+        }
+
         const cart = await CartItem.deleteMany({ _id: { $in: itemsCart }}).lean();
-
         if (!cart) {
-            errorCommonResponse(res, 'remove cart fail');
+            return errorCommonResponse(res, 'remove cart fail');
         }
 
         successResponse(res, result);
